requirements.txt
python-telegram-bot==21.3
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
